requirements.txt
flask
line-bot-sdk
schedule
gunicorn
